% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_edish.R
\name{edish_server}
\alias{edish_server}
\title{Server of the \code{dv.edish} module}
\usage{
edish_server(
  module_id,
  dataset_list,
  lb_date_var,
  subjectid_var = "USUBJID",
  arm_var = "ACTARM",
  visit_var = "VISIT",
  baseline_visit_val = "VISIT 01",
  lb_test_var = "LBTEST",
  at_choices = NULL,
  alp_choice = NULL,
  lb_result_var = "LBSTRESN",
  ref_range_upper_lim_var = "LBSTNRHI",
  norm_ref_lines = NULL,
  abs_ref_lines = NULL,
  uln_multiples = NULL,
  on_sbj_click = NULL
)
}
\arguments{
\item{module_id}{\verb{[character(1)]}

A unique ID string to create a namespace. Must match the ID of \code{edish_UI()}.}

\item{dataset_list}{\verb{[shiny::reactive(list(data.frame))]}

A reactive list of named datasets.}

\item{lb_date_var}{\verb{[character(1)]}

Name of the variable (\code{Date} or \code{POSIXt} class) containing the laboratory test date.}

\item{subjectid_var}{\verb{[character(1)]}

Name of the variable containing the unique subject IDs.}

\item{arm_var}{\verb{[character(1)]}

Name of the variable containing the arm/treatment information.}

\item{visit_var}{\verb{[character(1)]}

Name of the variable containing the visit information.}

\item{baseline_visit_val}{\verb{[character(1)]}

Character indicating which visit should be used as baseline visit.}

\item{lb_test_var}{\verb{[character(1)]}

Name of the variable containing the laboratory test information.}

\item{at_choices}{\verb{[character(1+)]}

Character vector specifying the possible choices of the x-axis aminotransferase laboratory test.}

\item{alp_choice}{\verb{[character(1) | NULL]}

Character vector specifying the alkaline phosphatase laboratory test choice.}

\item{lb_result_var}{\verb{[character(1)]}

Name of the variable containing results of the laboratory test.}

\item{ref_range_upper_lim_var}{\verb{[character(1)]}

Name of the variable containing the reference range upper limits.}

\item{norm_ref_lines}{\verb{[numeric(1+) | NULL]}

A named numeric vector of reference line values corresponding to normalized value laboratory
tests. Each value should be named with a value from \code{at_choices} or \code{tbili_choices}.}

\item{abs_ref_lines}{\verb{[numeric(1+) | NULL]}

A named numeric vector of reference line values corresponding to absolute value laboratory
tests. Each value should be named with a value from \code{at_choices} or \code{tbili_choices}.}

\item{uln_multiples}{\verb{[numeric(1+) | NULL]}

A named numeric vector of ULN multiples corresponding to normalized value laboratory
tests. Each value should be named with a value from \code{at_choices}.}

\item{on_sbj_click}{\verb{[function() | NULL]}

Function to invoke when a subject is clicked on the plot. If \code{NULL}, no action is taken.}
}
\value{
A reactive value containing the list of subjects in the clicked point, if applicable.
}
\description{
\code{edish_server()} contains the server of the \code{dv.edish} module.
}
\seealso{
\code{\link[=mod_edish]{mod_edish()}} and \code{\link[=edish_UI]{edish_UI()}}
}
