% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{filter_data}
\alias{filter_data}
\title{Filter data prior to plotting}
\usage{
filter_data(
  dataset,
  norm_ref_type,
  arm_var,
  sel_arm,
  lb_test_var,
  sel_lb_test,
  base_incl,
  uln_multiple
)
}
\arguments{
\item{dataset}{\verb{[data.frame]}

A data frame containing the columns specified by \code{lb_test_var} and \code{arm_var}.}

\item{norm_ref_type}{\verb{[character(1)]}

String indicating normalization reference type, either \code{"ULN"} or \code{"Baseline"}.}

\item{arm_var}{\verb{[character(1)]}

Name of the variable containing the arm/treatment information.}

\item{sel_arm}{\verb{[character(1+)]}

Character vector specifying a selection of arms/treatments.}

\item{lb_test_var}{\verb{[character(1)]}

Name of the variable containing the laboratory test parameter.}

\item{sel_lb_test}{\verb{[character(1)]}

String specifying a selected aminotransferase laboratory test.}

\item{base_incl}{\verb{[character(1)]}

String specifying the selected baseline inclusion choice, either \code{"LT"} (baseline less than threshold),
\code{"LTE"} (baseline less than equal to threshold), \code{"GTE"} (baseline greater than equal to threshold),
\code{"GT"} (baseline greater than threshold), or \code{"ALL"} (all).}

\item{uln_multiple}{\verb{[numeric(1)]}

Numeric ULN multiple used as threshold for baseline inclusions.}
}
\value{
A data frame.
}
\description{
Filter data on selected normalization reference type, arms/treatments and
aminotransferase laboratory test.
}
\keyword{internal}
