% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{generate_table}
\alias{generate_table}
\title{Generate a table of subject counts and percentages in areas delimited by reference lines}
\usage{
generate_table(
  dataset,
  subjectid_var,
  sel_x,
  sel_y,
  x_abs,
  y_abs,
  x_ref_line_num,
  y_ref_line_num
)
}
\arguments{
\item{dataset}{\verb{[data.frame]}

A data frame containing the variables listed below as columns.}

\item{subjectid_var}{\verb{[character(1)]}

Name of the variable containing the unique subject IDs.}

\item{sel_x}{\verb{[character(1)]}

String specifying the laboratory test to be displayed on the x-axis.}

\item{sel_y}{\verb{[character(1)]}

String specifying the laboratory test to be displayed on the y-axis.}

\item{x_abs}{\verb{[logical(1)]}

Logical indicating if absolute value should be plotted on x-axis.}

\item{y_abs}{\verb{[logical(1)]}

Logical indicating if absolute value should be plotted on y-axis.}

\item{x_ref_line_num}{\verb{[numeric(1)]}

Numeric specifying the reference line for the x-axis.}

\item{y_ref_line_num}{\verb{[numeric(1)]}

Numeric specifying the reference line for the y-axis.}
}
\value{
A data frame of subject counts and percentages categorized by normal/elevated laboratory tests.
}
\description{
Generate a table of subject counts and percentages in areas delimited by reference lines
}
\keyword{internal}
