% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mock_edish.R
\name{mock_edish_app}
\alias{mock_edish_app}
\title{Plain mock app without specific layout}
\usage{
mock_edish_app()
}
\description{
\code{mock_edish_app()} launches a mock app containing the dv.edish shiny module.
}
