% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{prepare_initial_data}
\alias{prepare_initial_data}
\title{Prepare initial data}
\usage{
prepare_initial_data(
  dataset_list,
  subjectid_var,
  arm_var,
  visit_var,
  lb_test_var,
  at_choices,
  tbili_choice,
  alp_choice,
  lb_date_var,
  lb_result_var,
  ref_range_upper_lim_var
)
}
\arguments{
\item{dataset_list}{\verb{[list(data.frame))]}

A list of datasets, containing a Demographics and a Lab Value dataset.

[list(data.frame))]: R:list(data.frame))}

\item{subjectid_var}{\verb{[character(1)]}

Name of the variable containing the unique subject IDs.}

\item{arm_var}{\verb{[character(1)]}

Name of the variable containing the arm/treatment information.}

\item{visit_var}{\verb{[character(1)]}

Name of the variable containing the visit information.}

\item{lb_test_var}{\verb{[character(1)]}

Name of the variable containing the laboratory test parameter.}

\item{at_choices}{\verb{[character(1+)]}

Character vector specifying the possible choices of the aminotransferase
laboratory test parameter for the x-axis.}

\item{tbili_choice}{\verb{[character(1)]}

Character vector specifying the choice of the total bilirubin laboratory
test parameter for the y-axis.}

\item{alp_choice}{\verb{[character(1) | NULL]}

Character vector specifying the choice of the alkaline phosphatase
laboratory test parameter.}

\item{lb_date_var}{\verb{[character(1)]}

Name of the variable (\code{Date} or \code{POSIXt} class) containing the laboratory test date.}

\item{lb_result_var}{\verb{[character(1)]}

Name of the variable containing results of the laboratory test.}

\item{ref_range_upper_lim_var}{\verb{[character(1)]}

Name of the variable containing the reference range upper limits.}
}
\value{
A data frame with the variables specified by the function arguments,
\code{subjectid_var}, \code{arm_var}, \code{visit_var}, \code{lb_test_var}, \code{lb_date_var},
\code{lb_result_var} and \code{ref_range_upper_lim_var}. Only rows for laboratory test
parameters specified by \code{at_choices}, \code{tbili_choice} and \code{alp_choice} are
kept.
}
\description{
Join DM and LB datasets into one, and filter on specified laboratory test
parameters, and maximum values for each subject/arm/test/visit group.
}
\keyword{internal}
