% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{derive_req_vars}
\alias{derive_req_vars}
\title{Derive required variables for plotting}
\usage{
derive_req_vars(
  dataset,
  subjectid_var,
  arm_var,
  visit_var,
  baseline_visit_val,
  lb_test_var,
  at_choices,
  tbili_choice,
  norm_ref_type,
  alp_choice,
  lb_date_var,
  lb_result_var,
  ref_range_upper_lim_var,
  by_visit,
  window_days
)
}
\arguments{
\item{dataset}{\verb{[data.frame]}

A data frame containing the data from \code{prepare_initial_data()}.}

\item{subjectid_var}{\verb{[character(1)]}

Name of the variable containing the unique subject IDs.}

\item{arm_var}{\verb{[character(1)]}

Name of the variable containing the arm/treatment information.}

\item{visit_var}{\verb{[character(1)]}

Name of the variable containing the visit information.}

\item{baseline_visit_val}{\verb{[character(1)]}

String indicating which visit should be used as baseline visit.}

\item{lb_test_var}{\verb{[character(1)]}

Name of the variable containing the laboratory test parameter.}

\item{at_choices}{\verb{[character(1+)]}

Character vector specifying the possible choices of the aminotransferase
laboratory test parameter for the x-axis.}

\item{tbili_choice}{\verb{[character(1)]}

Character vector specifying the choice of the total bilirubin laboratory
test parameter for the y-axis.}

\item{norm_ref_type}{\verb{[character(1)]}

String indicating normalization reference type, either \code{"ULN"} or \code{"Baseline"}.}

\item{alp_choice}{\verb{[character(1) | NULL]}

Character vector specifying the choice of the alkaline phosphatase
laboratory test parameter.}

\item{lb_date_var}{\verb{[character(1)]}

Name of the variable (\code{Date} or \code{POSIXt} class) containing the laboratory test date.}

\item{lb_result_var}{\verb{[character(1)]}

Name of the variable containing results of the laboratory test.}

\item{ref_range_upper_lim_var}{\verb{[character(1)]}

Name of the variable containing the reference range upper limits.}

\item{by_visit}{\verb{[logical(1)]}

A flag to indicate whether to plot aminotransferase values for each visit.}

\item{window_days}{\verb{[integer(1)]}

Window of the number of days considered between peaks.}
}
\value{
A data frame with the following derived variables:
\itemize{
\item \code{.ref_val}: ULN or baseline as the reference value for normalization.
\item \code{.visit_at}: Visit of peak aminotransferase value.
\item \code{.date_at}: Date of peak aminotransferase value.
\item \code{.norm_at}: Normalized peak aminotransferase value.
\item \code{.visit_tbili}: Visit of peak total bilirubin value.
\item \code{.date_tbili}: Date of peak total bilirubin value.
\item \code{.norm_tbili}: Normalized peak total bilirubin value.
\item \code{.norm_alp}: Normalized alkaline phosphotase value at same visit as aminotransferase value.
\item \code{.offset_days}: Number of days from aminotransferase visit to total bilirubin visit.
\item \code{.norm_ref_type}: Normalization reference type, copied from \code{norm_ref_type} argument.
}
}
\description{
Derive required variables for plotting
}
\keyword{internal}
