% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{derive_req_vars}
\alias{derive_req_vars}
\title{Derive required variables}
\usage{
derive_req_vars(
  dataset,
  subjectid_var,
  arm_var,
  visit_var,
  lb_test_var,
  lb_result_var,
  ref_range_upper_lim_var,
  sel_x,
  sel_y
)
}
\arguments{
\item{dataset}{\verb{[data.frame]}

A dataframe containing the variables listed below as columns.}

\item{subjectid_var}{\verb{[character(1)]}

Name of the variable containing the unique subject IDs.}

\item{arm_var}{\verb{[character(1)]}

Name of the variable containing the arm/treatment information.}

\item{visit_var}{\verb{[character(1)]}

Name of the variable containing the visit information.}

\item{lb_test_var}{\verb{[character(1)]}

Name of the variable containing the laboratory test information.}

\item{lb_result_var}{\verb{[character(1)]}

Name of the variable containing results of the laboratory test.}

\item{ref_range_upper_lim_var}{\verb{[character(1)]}

Name of the variable containing the reference range upper limits.}

\item{sel_x}{\verb{[character(1)]}

Character specifying the laboratory test selected for the x-axis.}

\item{sel_y}{\verb{[character(1)]}

Character specifying the laboratory test selected for the y-axis.}
}
\value{
The restructured dataset.
}
\description{
\code{derive_req_vars()} restructures the stated dataset to include variables containing
the ratio of a lab result divided by ULN or the baseline value. The corresponding variable
names are shaped as follows: "r_\if{html}{\out{<ULN or Baseline>}}_\if{html}{\out{<selected lab test>}}.
}
\keyword{internal}
