% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_edish.R
\name{edish_server}
\alias{edish_server}
\title{Server of the \code{dv.edish} module}
\usage{
edish_server(
  module_id,
  dataset_list,
  subjectid_var = "USUBJID",
  arm_var = "ACTARM",
  arm_default_vals = NULL,
  visit_var = "VISIT",
  baseline_visit_val = "VISIT 01",
  lb_test_var = "LBTEST",
  lb_test_choices = c("Alkaline Phosphatase", "Alanine Aminotransferase",
    "Aspartate Aminotransferase", "Bilirubin"),
  lb_test_default_x_val = "Aspartate Aminotransferase",
  lb_test_default_y_val = "Bilirubin",
  lb_result_var = "LBSTRESN",
  ref_range_upper_lim_var = "LBSTNRHI",
  receiver_id = NULL,
  afmm_param = NULL
)
}
\arguments{
\item{module_id}{\verb{[character(1)]}

A unique ID string to create a namespace. Must match the ID of \code{edish_UI()}.}

\item{dataset_list}{\verb{[shiny::reactive(list(data.frame))]}

A reactive list of named datasets.}

\item{subjectid_var}{\verb{[character(1)]}

Name of the variable containing the unique subject IDs. Defaults to \code{"USUBJID"}.}

\item{arm_var}{\verb{[character(1)]}

Name of the variable containing the arm/treatment information. Defaults to \code{"ACTARM"}.}

\item{arm_default_vals}{\verb{[character(1+)]}

Vector specifying the default value(s) for the arm selector. Defaults to \code{NULL}.}

\item{visit_var}{\verb{[character(1)]}

Name of the variable containing the visit information. Defaults to \code{"VISIT"}.}

\item{baseline_visit_val}{\verb{[character(1)]}

Character indicating which visit should be used as baseline visit. Defaults to \code{"VISIT 01"}.}

\item{lb_test_var}{\verb{[character(1)]}

Name of the variable containing the laboratory test information. Defaults to \code{"LBTEST"}.}

\item{lb_test_choices}{\verb{[character(1+)]}

Character vector specifying the possible choices of the laboratory test. Defaults to
\code{c("Alkaline Phosphatase", "Alanine Aminotransferase", "Aspartate Aminotransferase", "Bilirubin")}}

\item{lb_test_default_x_val}{\verb{[character(1)]}

Character specifying the default laboratory test choice for the plot's x-axis.
Defaults to \code{"Aspartate Aminotransferase"}.}

\item{lb_test_default_y_val}{\verb{[character(1)]}

Character specifying the default laboratory test choice for the plot's y-axis.
Defaults to \code{"Bilirubin"}.}

\item{lb_result_var}{\verb{[character(1)]}

Name of the variable containing results of the laboratory test. Defaults to \code{"LBSTRESN"}.
In case of multiple values in \code{lb_result_var} per \code{subjectid_var}, \code{visit_var}, and
\code{lb_test_var}, only the maximum value will be used. Note that a NA value in the considered values
will cause a value of NA to be returned as maximum value.}

\item{ref_range_upper_lim_var}{\verb{[character(1)]}

Name of the variable containing the reference range upper limits.
Defaults to \code{"LBSTNRHI"}.}

\item{receiver_id}{\verb{[character(1) | NULL]}

Character string defining the ID of the module to which to send a subject ID. The
module must exist in the module list. The default is NULL which disables communication.}

\item{afmm_param}{\verb{[list]}

Named list of a selection of arguments from module manager. Expects
at least two elements: \code{utils} and \code{module_names} defining a character vector
whose entries have the corresponding module IDs as names.}
}
\description{
\code{edish_server()} contains the server of the \code{dv.edish} module.
}
\seealso{
\code{\link[=mod_edish]{mod_edish()}} and \code{\link[=edish_UI]{edish_UI()}}
}
