% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_edish.R
\name{edish_server}
\alias{edish_server}
\title{Server of the \code{dv.edish} module}
\usage{
edish_server(
  module_id,
  dataset_list,
  subjectid_var = "USUBJID",
  arm_var = "ACTARM",
  arm_default_vals = NULL,
  visit_var = "VISIT",
  baseline_visit_val = "VISIT 01",
  lb_test_var = "LBTEST",
  lb_test_choices = c("Alkaline Phosphatase", "Alanine Aminotransferase",
    "Aspartate Aminotransferase", "Bilirubin"),
  lb_test_default_x_val = "Aspartate Aminotransferase",
  lb_test_default_y_val = "Bilirubin",
  lb_result_var = "LBSTRESN",
  ref_range_upper_lim_var = "LBSTNRHI"
)
}
\arguments{
\item{module_id}{\verb{[character(1)]}

A unique ID string to create a namespace. Must match the ID of \code{edish_UI()}.}

\item{dataset_list}{\verb{[shiny::reactive(list(data.frame))]}

A reactive list of named datasets.}

\item{subjectid_var}{\verb{[character(1)]}

Name of the variable containing the unique subject IDs.}

\item{arm_var}{\verb{[character(1)]}

Name of the variable containing the arm/treatment information.}

\item{arm_default_vals}{\verb{[character(1+)]}

Vector specifying the default value(s) for the arm selector.}

\item{visit_var}{\verb{[character(1)]}

Name of the variable containing the visit information.}

\item{baseline_visit_val}{\verb{[character(1)]}

Character indicating which visit should be used as baseline visit.}

\item{lb_test_var}{\verb{[character(1)]}

Name of the variable containing the laboratory test information.}

\item{lb_test_choices}{\verb{[character(1+)]}

Character vector specifying the possible choices of the laboratory test.}

\item{lb_test_default_x_val}{\verb{[character(1)]}

Character specifying the default laboratory test choice for the plot's x-axis.}

\item{lb_test_default_y_val}{\verb{[character(1)]}

Character specifying the default laboratory test choice for the plot's y-axis.}

\item{lb_result_var}{\verb{[character(1)]}

Name of the variable containing results of the laboratory test.}

\item{ref_range_upper_lim_var}{\verb{[character(1)]}

Name of the variable containing the reference range upper limits.}
}
\description{
\code{edish_server()} contains the server of the \code{dv.edish} module.
}
\seealso{
\code{\link[=mod_edish]{mod_edish()}} and \code{\link[=edish_UI]{edish_UI()}}
}
