% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_edish.R
\name{edish_UI}
\alias{edish_UI}
\title{User Interface of the \code{dv.edish} module}
\usage{
edish_UI(
  module_id,
  arm_default_vals,
  at_choices,
  at_default_val,
  tbili_choice,
  default_by_visit,
  window_days
)
}
\arguments{
\item{module_id}{\verb{[character(1)]}

A unique ID string to create a namespace. Must match the ID of \code{edish_server()}.}

\item{arm_default_vals}{\verb{[character(1+)]}

Vector specifying the default value(s) for the arm selector. Defaults to \code{NULL}.}

\item{at_choices}{\verb{[character(1+)]}

Character vector specifying the possible choices of the x-axis aminotransferase laboratory test.}

\item{at_default_val}{\verb{[character(1)]}

Character specifying the default x-axis aminotransferase laboratory test choice.}

\item{tbili_choice}{\verb{[character(1)]}

Character vector specifying the y-axis total bilirubin laboratory test choice.}

\item{default_by_visit}{\verb{[logical(1)]}

A flag to indicate the default of whether or not to plot aminotransferase values for each visit.}

\item{window_days}{\verb{[integer(1) | NULL]}

Window of the number of days considered between peaks.}
}
\value{
A shiny \code{uiOutput} element.
}
\description{
\code{edish_UI()} contains the UI of the \code{dv.edish} module.
}
\seealso{
\code{\link[=mod_edish]{mod_edish()}} and \code{\link[=edish_server]{edish_server()}}
}
