% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_edish.R
\name{edish_UI}
\alias{edish_UI}
\title{User Interface of the \code{dv.edish} module}
\usage{
edish_UI(module_id)
}
\arguments{
\item{module_id}{\verb{[character(1)]}

A unique ID string to create a namespace. Must match the ID of \code{edish_server()}.}
}
\value{
A shiny \code{uiOutput} element.
}
\description{
\code{edish_UI()} contains the UI of the \code{dv.edish} module.
}
\seealso{
\code{\link[=mod_edish]{mod_edish()}} and \code{\link[=edish_server]{edish_server()}}
}
