% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{filter_data}
\alias{filter_data}
\title{Filter data prior to plotting}
\usage{
filter_data(dataset, norm_ref_type, arm_var, sel_arm, lb_test_var, sel_lb_test)
}
\arguments{
\item{dataset}{\verb{[data.frame]}

A data frame containing the columns specified by \code{lb_test_var} and \code{arm_var}.}

\item{norm_ref_type}{\verb{[character(1)]}

String indicating normalization reference type, either \code{"ULN"} or \code{"Baseline"}.}

\item{arm_var}{\verb{[character(1)]}

Name of the variable containing the arm/treatment information.}

\item{sel_arm}{\verb{[character(1+)]}

Character vector specifying a selection of arms/treatments.}

\item{lb_test_var}{\verb{[character(1)]}

Name of the variable containing the laboratory test parameter.}

\item{sel_lb_test}{\verb{[character(1)]}

Character vector specifying a selected aminotransferase laboratory test.}
}
\value{
A data frame.
}
\description{
Filter data on selected normalization reference type, arms/treatments and
aminotransferase laboratory test.
}
\keyword{internal}
