% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{filter_data}
\alias{filter_data}
\title{Filter data}
\usage{
filter_data(dataset, arm_var, sel_arm, lb_test_var, sel_lb_test)
}
\arguments{
\item{dataset}{\verb{[data.frame]}

A dataframe containing the columns specified by \code{lb_test_var} and \code{arm_var}.}

\item{arm_var}{\verb{[character(1)]}

Name of the variable containing the arm/treatment information.}

\item{sel_arm}{\verb{[character(1+)]}

Character vector specifying a selection of arms/treatments.}

\item{lb_test_var}{\verb{[character(1)]}

Name of the variable containing the laboratory test information.}

\item{sel_lb_test}{\verb{[character(1+)]}

Character vector specifying a selection of laboratory tests.}
}
\value{
The filtered dataset.
}
\description{
\code{filter_data()} filters \code{dataset} to only contain the values of \code{sel_lb_test}
in the \code{lb_test_var} column and the values of \code{sel_arm} in the \code{arm_var} column.
}
\keyword{internal}
