% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{generate_plot}
\alias{generate_plot}
\title{Generate an eDISH plot}
\usage{
generate_plot(
  dataset,
  subjectid_var,
  arm_var,
  sel_x,
  sel_y,
  norm_ref_type,
  x_ref_line_num,
  y_ref_line_num,
  x_rng_lower,
  x_rng_upper,
  y_rng_lower,
  y_rng_upper,
  alp_flag,
  by_visit
)
}
\arguments{
\item{dataset}{\verb{[data.frame]}

A data frame containing the variables listed below as columns.}

\item{subjectid_var}{\verb{[character(1)]}

Name of the variable containing the unique subject IDs.}

\item{arm_var}{\verb{[character(1)]}

Name of the variable containing the arm/treatment information.}

\item{sel_x}{\verb{[character(1)]}

String specifying the laboratory test to be displayed on the x-axis.}

\item{sel_y}{\verb{[character(1)]}

String specifying the laboratory test to be displayed on the y-axis.}

\item{norm_ref_type}{\verb{[character(1)]}

String indicating normalization reference type, either \code{"ULN"} or \code{"Baseline"}.}

\item{x_ref_line_num}{\verb{[numeric(1)]}

Numeric specifying the reference line for the x-axis.}

\item{y_ref_line_num}{\verb{[numeric(1)]}

Numeric specifying the reference line for the y-axis.}

\item{x_rng_lower}{\verb{[numeric(1) | NULL]}

Numeric specifying the lower limit in the x-axis range.}

\item{x_rng_upper}{\verb{[numeric(1) | NULL]}

Numeric specifying the upper limit in the x-axis range.}

\item{y_rng_lower}{\verb{[numeric(1) | NULL]}

Numeric specifying the lower limit in the y-axis range.}

\item{y_rng_upper}{\verb{[numeric(1) | NULL]}

Numeric specifying the upper limit in the y-axis range.}

\item{alp_flag}{\verb{[logical(1)]}

Logical indicating if ALP data was requested.}

\item{by_visit}{\verb{[logical(1)]}

A flag to indicate whether to plot aminotransferase values for each visit.}

\item{visit_var}{\verb{[character(1)]}

Name of the variable containing the visit information.}
}
\value{
An object specifying the generated eDISH plot.
}
\description{
Generate an eDISH plot
}
\keyword{internal}
