% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{generate_plot}
\alias{generate_plot}
\title{Generate plot}
\usage{
generate_plot(
  dataset,
  subjectid_var,
  arm_var,
  visit_var,
  sel_x,
  sel_y,
  x_plot_type,
  y_plot_type,
  x_ref_line_num,
  y_ref_line_num
)
}
\arguments{
\item{dataset}{\verb{[data.frame]}

A dataframe containing the variables listed below as columns.}

\item{subjectid_var}{\verb{[character(1)]}

Name of the variable containing the unique subject IDs.}

\item{arm_var}{\verb{[character(1)]}

Name of the variable containing the arm/treatment information.}

\item{visit_var}{\verb{[character(1)]}

Name of the variable containing the visit information.}

\item{sel_x}{\verb{[character(1)]}

Character specifying the laboratory test to be displayed on the x-axis.}

\item{sel_y}{\verb{[character(1)]}

Character specifying the laboratory test to be displayed on the y-axis.}

\item{x_plot_type}{\verb{[character(1)]}

Character specifying the plot type for the x-axis. This leads to
using the \code{dataset}'s column "r_<x_plot_type>_<x_sel>" for the x-values.}

\item{y_plot_type}{\verb{[character(1)]}

Character specifying the plot type for the y-axis. This leads to
using the \code{dataset}'s column "r_<y_plot_type>_<y_sel>" for the y-values.}
}
\value{
A plotly object specifying the generated eDISH plot.
}
\description{
\code{generate_plot()} generates an eDISH plot by means of the \pkg{plotly} package.
}
\keyword{internal}
