% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mock_edish.R
\name{mock_edish_mm}
\alias{mock_edish_mm}
\title{Mock app integrated in the module manager framework}
\usage{
mock_edish_mm()
}
\description{
\code{mock_table_mm()} launches a mock app containing the dv.edish shiny module by means of the \code{dv.manager}.
}
\keyword{internal}
