% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_edish.R
\name{mod_edish}
\alias{mod_edish}
\title{eDISH module}
\usage{
mod_edish(
  module_id,
  subject_level_dataset_name,
  lab_dataset_name,
  lb_date_var,
  subjectid_var = "USUBJID",
  arm_var = "ACTARM",
  arm_default_vals = NULL,
  visit_var = "VISIT",
  baseline_visit_val = "VISIT 01",
  lb_test_var = "LBTEST",
  at_choices = c("Alanine Aminotransferase", "Aspartate Aminotransferase"),
  at_default_val = "Aspartate Aminotransferase",
  tbili_choices = "Bilirubin",
  tbili_default_val = "Bilirubin",
  alp_choice = "Alkaline Phosphatase",
  lb_result_var = "LBSTRESN",
  ref_range_upper_lim_var = "LBSTNRHI",
  default_by_visit = FALSE,
  window_days = NULL,
  receiver_id = NULL
)
}
\arguments{
\item{module_id}{\verb{[character(1)]}

A unique module ID.}

\item{subject_level_dataset_name}{\verb{[character(1)]}

Name of the subject-level dataset to be used.}

\item{lab_dataset_name}{\verb{[character(1)]}

Name of the laboratory results dataset to be used.}

\item{lb_date_var}{\verb{[character(1)]}

Name of the variable (\code{Date} or \code{POSIXt} class) containing the laboratory test date.}

\item{subjectid_var}{\verb{[character(1)]}

Name of the variable containing the unique subject IDs.}

\item{arm_var}{\verb{[character(1)]}

Name of the variable containing the arm/treatment information.}

\item{arm_default_vals}{\verb{[character(1+) | NULL]}

Vector specifying the default value(s) for the arm selector.}

\item{visit_var}{\verb{[character(1)]}

Name of the variable containing the visit information.}

\item{baseline_visit_val}{\verb{[character(1)]}

Character indicating which visit should be used as baseline visit.}

\item{lb_test_var}{\verb{[character(1)]}

Name of the variable containing the laboratory test information.}

\item{at_choices}{\verb{[character(1+)]}

Character vector specifying the possible choices of the x-axis aminotransferase laboratory test.}

\item{at_default_val}{\verb{[character(1) | NULL]}

Character specifying the default x-axis aminotransferase laboratory test choice.}

\item{tbili_choices}{\verb{[character(1+)]}

Character vector specifying the possible choices of the y-axis liver function laboratory test
(e.g. "Bilirubin", "Prothrombin Intl. Normalized Ratio", etc.).}

\item{tbili_default_val}{\verb{[character(1) | NULL]}

Character specifying the default y-axis liver function laboratory test choice.}

\item{alp_choice}{\verb{[character(1) | NULL]}

Character vector specifying the alkaline phosphatase laboratory test choice.}

\item{lb_result_var}{\verb{[character(1)]}

Name of the variable containing results of the laboratory test.}

\item{ref_range_upper_lim_var}{\verb{[character(1)]}

Name of the variable containing the reference range upper limits.}

\item{default_by_visit}{\verb{[logical(1)]}

A flag to indicate the default of whether or not to plot aminotransferase values for each visit.}

\item{window_days}{\verb{[integer(1) | NULL]}

Window of the number of days considered between peaks.}

\item{receiver_id}{\verb{[character(1) | NULL]}

Character string defining the ID of the module to which to send a subject ID. The
module must exist in the module list. The default is NULL which disables communication.}
}
\value{
A list containing the following elements to be used by the
\pkg{dv.manager}:
\itemize{
\item{\code{ui}: A UI function of the \pkg{dv.edish} module.}
\item{\code{server}: A server function of the \pkg{dv.edish} module.}
\item{\code{module_id}: A unique identifier.}
}
}
\description{
\code{mod_edish()} displays the (modified) evaluation of Drug-Induced Serious Hepatotoxicity (eDISH/mDISH) plot
to support the assessment of drug-induced liver injury (DILI). The scatter plot depicts the correlation
between peak values of an aminotransferase parameter and a liver function parameter on a subject level.
}
