% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{prepare_initial_data}
\alias{prepare_initial_data}
\title{Prepare initial data}
\usage{
prepare_initial_data(
  dataset_list,
  subjectid_var,
  arm_var,
  visit_var,
  baseline_visit_val,
  lb_test_var,
  lb_test_choices,
  lb_result_var,
  ref_range_upper_lim_var
)
}
\arguments{
\item{dataset_list}{\verb{[list(data.frame))]}

A list of datasets, containing a Demographics and a Lab Value dataset.

[list(data.frame))]: R:list(data.frame))}

\item{subjectid_var}{\verb{[character(1)]}

Name of the variable containing the unique subject IDs.}

\item{arm_var}{\verb{[character(1)]}

Name of the variable containing the arm/treatment information.}

\item{visit_var}{\verb{[character(1)]}

Name of the variable containing the visit information.}

\item{baseline_visit_val}{\verb{[character(1)]}

Character indicating which visit should be used as baseline visit.}

\item{lb_test_var}{\verb{[character(1)]}

Name of the variable containing the laboratory test information.}

\item{lb_test_choices}{\verb{[character(1+)]}

Character vector specifying the possible choices of the laboratory test.}

\item{lb_result_var}{\verb{[character(1)]}

Name of the variable containing results of the laboratory test.}

\item{ref_range_upper_lim_var}{\verb{[character(1)]}

Name of the variable containing the reference range upper limits.}
}
\value{
A single dataframe including columns defined by \code{subjectid_var},
\code{arm_var}, \code{visit_var}, \code{lb_test_var}, \code{lb_result_var}, and \code{ref_range_upper_lim_var},
as well as the column "BASE" containing the corresponding baseline values.
In case of multiple values in \code{lb_result_var} per \code{subjectid_var}, \code{visit_var}, and
\code{lb_test_var}, only the maximum value will be used. Note that a NA value in the considered values
will cause a value of NA to be returned as maximum value.
}
\description{
\code{prepare_initial_data()} prepares data initially by restructuring
and joining DM and LB dataset into one.
}
\keyword{internal}
